GUI code added
There is now a file that can be run through shoes to produce the GUI.
The gui has mulitple fields and will run the trainAndValidate code with
the given arguments. Important updates:
1) There is framework in teh GUI to add more transition functions and I
tested that it will work using the alerts (which we can remove, they
were just to see that I made it to the branches). We just need to add
more transition functions and a method that allows you to choose your
transition function and run the algorithm.
2) We can use the golddigger code, it won't be hard to change I just
didn't want to add a loop field if we didn't want it in the final gui
and didn't want to set a default value without discussing it.
3) I'm slighlty concerned that the naive were returning 0.23 and the
network was returning 11 -70 depending on the arguments I gave it. Do we
know why?